get country from list of country codes and navigate to category page to query recipe API
This change gets the country code from the coords recieved and sent to the geolocation API and
compares the result to a local JSON file to get the name of the country. The name of the country
is then sent to the recipe API to display the recipes native to that country.